Перепись loadStrategy падение по rateLimit

+++ etl-service/src/main/java/com/sitefilm/etl/infrastructure/provider/tmdb/response/CountPageResponse.java
@@ -4,6 +4,6 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 
 public record CountPageResponse(
         @JsonProperty("total_pages")
-        Integer total_pages
+        Short total_pages
 ) {
 }
